+++ api/__init__.py
@@ -1,11 +1,13 @@
-import logging
 from flask import Flask
 from flask_cors import CORS
 from celery_app import init_celery
 from database.extensions import db, migrate
+from logging_config.config import setup_logging
 
 
 def create_app():
+    setup_logging()
+
     app = Flask(__name__)
 
     CORS(app, origins=[
@@ -14,7 +16,6 @@ def create_app():
     ])
 
     app.config.from_object("config.Config")
-    app.logger.setLevel(logging.INFO)
 
     db.init_app(app)
     migrate.init_app(app, db, directory="database/migrations")